requirements.txt
beautifulsoup4==4.12.2
Django==4.2.3
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
rembg==2.0.50
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tensorboard==2.10.1
tensorflow-gpu==2.10.0
urllib3==1.26.16